requirements.txt
discord.py[voice]
python-dotenv
pyttsx3
:t_dlp
